requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
deep-translator==1.11.4
gTTS==2.5.4
SQLAlchemy==2.0.49
psycopg[binary]
python-dotenv==1.2.2
python-jose[cryptography]
bcrypt==5.0.0
passlib==1.7.4
pydantic[email]==2.12.5